Add all jenkins jobs to setup.py
* Archive unused scripts
* Add default config
