+++ quasar/mobile_commons.py
@@ -148,3 +148,33 @@ def scrape_campaigns():
 
     cur.close()
     db.close()
+
+def convert_campaign_lookup_to_id():
+    db, cur = connect_db()
+    cur.execute("SELECT * FROM users_and_activities.mobile_campaign_id_lookup")
+    result = cur.fetchall()
+
+    for row in result:
+        if row['nid'] is not None:
+            web_alpha = "0"
+            if "alpha" in row['opt_in_path_id_name'] or "Alpha" in row['opt_in_path_id_name']:
+                web_alpha = "1"
+
+            nid = row['nid']
+            campaign_id = row['mobile_campaign_id']
+            opt_in_id = row['opt_in_path_id']
+            web_alpha = "1"
+            if row['run_nid'] is None:
+                campaign_run = "NULL"
+                insert_campaign = "insert ignore into users_and_activities.mobile_campaign_ids VALUES({0}, {1}, {2}, {3}, \"{4}\")".format(
+                    nid, campaign_id, opt_in_id, web_alpha, campaign_run)
+            else:
+                campaign_run = row['run_nid']
+                insert_campaign = "insert ignore into users_and_activities.mobile_campaign_ids VALUES({0}, {1}, {2}, {3}, {4})".format(
+                    nid, campaign_id, opt_in_id, web_alpha, campaign_run)
+            print(insert_campaign)
+            cur.execute(insert_campaign)
+            db.commit()
+
+    cur.close()
+    db.close()
