+++ quasar/unused/MailchimpWebScraper.py
