+++ quasar/unused/mailchimp_international_subscriber_to_user_table.py
