+++ quasar/unused/mailchimp_subscribers_daily_snapshot.py
